requirements.txt
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.11.2
mysqlclient==2.1.0
PyMySQL==1.0.2
python-dotenv==0.20.0
pytz==2022.1
PyJWT==2.4.0
boto3==1.24.17